requirements.txt
bcrypt==4.3.0
blinker==1.9.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.1
et_xmlfile==2.0.0
Flask==3.1.2
invoke==2.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.2
paramiko==4.0.0
pycparser==2.23
PyNaCl==1.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3
